requirements.txt
langchain>=0.1.0
faiss-cpu>=1.7.4
openai>=1.0.0
python-dotenv>=1.0.0
tqdm
unstructured
chromadb  # optional if switching vector store
